requirements.txt
discord.py
faiss-cpu
numpy
scikit-learn
asyncio